LayoutTests/imported/w3c/web-platform-tests/shadow-dom/declarative/tentative/shadowrootadoptedstylesheets/shadowrootadoptedstylesheets-async-fetch-expected.txt


FAIL Async-fetched CSS module replaces placeholder and applies styles. assert_equals: Before fetch: expected 1 placeholder(s). expected 1 but got 0
